Write Erlang cookie to both vcap & root home dirs
Even though the _ARGS environment variables work fine in 3.6, the new,
Elixir-based rabbitmqctl in 3.7 does not respect the -setcookie flag.
It's escript, and -setcookie is only recognised by erl(1) or erlc(1).

Furthermore, the native clustering doesn't respect the -setcookie flag.
While we have confirmed it as a bug and will be addressing it soon, the
.erlang.cookie file is the most reliable way of setting the Erlang
cookie. We are losing the ability to run rabbitmqctl with the temporary
system user that BOSH creates (e.g. bosh_), but arguably we should never
run rabbitmqctl commands as this user anyway.