Do not hard kill the Erlang VM if it fails to stop cleanly
Having discussed with @dumbbell, we should never hard kill the Erlang
VM. Based on its state, only an operator can take this decision. We
agree that we should improve this in RabbitMQ itself, but until we can
reproduce some of the edge cases seen in the field, we can't make those
improvements.

[#142938685]

Signed-off-by: Gerhard Lazu <gerhard@rabbitmq.com>