Use the safest disk_free_limit production default
If RabbitMQ is using all available memory and needs to flush to disk, as
would be the case when shutting down, there should be sufficient space
left on disk for RabbitMQ to start **after** it successfully stops.

RabbitMQ, like any stateful service (e.g. PostgreSQL), crashes hard when
it runs out of disk space. Furthermore, if there is no disk space
avaialble, it will fail to start.

Let's be as defensive as possible and stop all publishers when there is
less than 2 x RAM disk space available.