Add bindings app
It was useful to determine whether the scheduble bind type makes a
difference when sharing CPUs with other competing processes. It didn't
make a difference in this case, but it can make a difference when using
something more memory & CPU-intensive, such as Java.

This is an antipattern (run apps in CF), I'm only committing this in
case we need it in the future (will revert in next commit).

[#148072315]