Fold pre-start and post-start into start
We do not know when pre-start and post-start
will run so we have to enable plugins right
before (or after) starting RabbitMQ.

While at it, enable all plugins without --offline
after the node is running but rabbitmq_management before
(with --offline).

In 3.7.0 rabbitmq-plugins will perform stricter and more
explicit execution environment validation and unless --offline
is provided it will require RabbitMQ app to be running on the
target node.

Because server download, installation and startup now can take longer
than 30 seconds, we bump Monit wait period to 90s.

This fixes plugin activation in the 3.7.0 LRE.

Signed-Off-By: Gerhard Lazu <glazu@pivotal.io>