Remove unused Erlang versions
Some older version are not self-contained, and they don't appear to be
as useful as the newer ones. If we need a specific Erlang version, such as
the latest 18.3, we can add it at the right moment. For now, we will
shorten our compilation phase when bumping the stemcell and only have
the Erlang versions which we need.