Add man pages
While at it, move all functions from start into commands. This allows us
to run ad-hoc commands individually, and makes it easy to invoke
arbitrary job logic from different BOSH job lifecycles, if we choose to
do so.

Drain & stop are next contenders for this type of refactoring. Since
they do not suffer from the same level of churn as start, they can stay
as they are for a while longer.

[closes #14]