Increase the max wait time for downloading generic unix packages
It was observed a number of times that packages can take longer than 45
seconds to download. This results in incomplete zip files which can be
unarchived but do not contain the full contents of the original,
including the `sbin` dir, which means that all subsequent steps will
fail.

This needs further refinement, but it improves things, so it's a good
step.