Add observer_cli [1] and integrate with RabbitMQ via rabbitmq-observe
ncurses-based interface for introspecting the Erlang VM running RabbitMQ
from the command-line - similar to htop for Linux.

Even though it requires an the user to type (copy/paste) and extra
command once the remote shell starts, it's a small price to pay for the
convenience that it provides. Look at entop [2] on how to improve this.
Thanks @ten0s!

Compiled against the oldest version of Erlang that is supported by
RabbitMQ 3.6 so that it works with any RabbitMQ & Erlang combination.

Added as an optional BOSH link, so that it leaves no trace if it's not
included.

Closes #42

[1] https://github.com/zhongwencool/observer_cli
[2] https://github.com/mazenharake/entop/blob/master/entop