Enable tcp_listen_options customisation
Tuning the TCP buffers can result in a significant memory reduction for
RabbitMQ nodes that service many connections.

For a RabbitMQ node with 20k connections that used to require 12GB of
RAM, after `{tcp_listen_options, [{sndbuf, 4096}, {recbuf, 4096}]` was
set, memory usage dropped to 5GB - a 58.4% reduction in memory
consumption. Please remember that this setting is given as an example,
and it may be counter-productive for your RabbitMQ workload.