Template release version, do not hard-code
This property was introduced not that long ago, it will not be available
in older directors. Since we are always on the latest stable version of
BOSH, 265.2.0 right now, this does not affect us.

One less-obvious fact about this version is that it uses the following
format: CURRENT_FINAL_VERSION+dev.UNIX_TIMESTAMP. I was personally
expecting the CURRENT_FINAL_VERSION part to be NEXT_FINAL_VERSION, but
since we don't what what the next release version will be, BOSH will use
`+` instead of `-` and the CURRENT_FINAL_VERSION. Neat semver detail!
Read more about `+` vs `-` in the context of semver: https://semver.org/#spec-item-10

Thanks @cppforlife!