Store generic-unix tgz on the data disk, at a fixed path
This saves on having to download it every single time the
rabbitmq-server job changes.

The package is still being copied and run from the job's packages. This
ensures that we discard all previous state when the rabbitmq-server job
changes, which is very useful for things such as plugin changes - no
plugins need to be disabled before only specific plugins are to be
enabled. This makes us stay true to BOSH's philosophy of idempotence and
declarative nature, we are only optimizing artefact caching.

[delivers #157095778]