Do not try to stop EPMD processes that are not ours
The assumption is that if there is an EMPD process that is not ours,
someone is doing something on purpose, and we should not interfere with
this EPMD instance. Besides, even if we tried to stop it, the PID that
we are tracking will not match, so the command will fail anyways.