Add BOSH job for RabbitMQ Perf Test
For a long time, we did the easiest thing which was to just push
RabbitMQ PerfTest to CloudFoundry. We've soon learned that it was easy
for a few PerfTest instances to start contending for CPU and/or network
which would skew the benchmark results. Even though CloudFoundry
isolation segments can mitigate against this, we've learned that
measuring message latency with PerfTest running within a container was
producing unreliable measurements due to the way TCP/IP packets are
routed, using iptables. Even though this was reflecting the CloudFoundry
reality, it wasn't a fair representation for RabbitMQ or PerfTest. For
more context, see
https://github.com/rabbitmq/workloads/tree/8c1b35585ac94c773f2e5afc3e5cbf9a887e7a69/low-latency

Lastly, we've come across environments where we don't have access to
CloudFoundry, only a BOSH Director. This was the tipping point which
made it obvious that we need isolated PerfTest environments, sitting
as close as possible to the broker, but not interfering with anything
else.