Do not enable rabbitmq_management by default
It should be possible to deploy RabbitMQ without rabbitmq_management, as
we've learned from deployments where performance or security are
important.

When we want to test the performance implications of running the
metrics subsystem in RabbitMQ, we want to disable rabbitmq_management
since the metrics subsystem is tied to rabbitmq_management (especially
rabbitmq_management_agent).