Update observer_cli to v1.3.3, now compatible with OTP 21
Updates recon to v2.3.6.

We're now prepending both observer_cli & recon to Erlang's load path so
that the latest versions get loaded. Even though RabbitMQ 3.7.8-rc.3
ships with recon v2.3.2, we're happy to run the latest and greatest. It
would be nice to contribute a PR in RabbitMQ itself once we are
confident that the new recon version works well.