Revert not listing local dev releases when deploying
When we need to test release changes, we use dev releases, so it makes
sense to list them when creating new deploys. Without this the workflow
requires the user to abort the deploy after the configuration is saved,
modify the configuration, then continue the deploy, which is just weird.

This basically goes back to
https://github.com/rabbitmq/rabbitmq-server-boshrelease/commit/2429eea778ab285263d4536d0be11621df08096a#commitcomment-29047610
and reverts
https://github.com/rabbitmq/rabbitmq-server-boshrelease/commit/222252b000d020e738bba963c8125c619e44070f

I've since upgraded to macOS High Sierra, so we're good : )