Explain tcpdump options, remove timeout feature
Some tcpdump flags do not have a long equivalent, such as -A or -w, so
we simply wrap them in function calls that capture the intent for using
those flags.

timeout returns a non-zero exit status if the command doesn't return
within the expected interval. Semantically, this means that the command
failed. In our case, tcpdump didn't fail, it simply was meant to capture
packets indefinitely.

We pass through all arguments, same as rabbitmq-admin, so that the
caller can modify the script's default behaviour (n.b. not all flags can
be overriden).

Notice that we load the env explicitly, and that we run in debug mode if
the rabbitmq-server.debug property has been set to true. Even though the
rabbitmq-server job defaults it to false,
rabbitmq-server-deployment-template.yml defaults it to false. All
deployments done using `gmake deploy` are still considered dev/debug/QA,
as per the original intent of this BOSH release.

Happy to discuss @MarcialRosales

[#160090949]