Change scripts so that they dont override DEBUG
We want to be able to call any script with the ability to
define if we want to have BASH debugging or not by declaring
the env var DEBUG with either true or false.

This change also refactors the logic that checks the DEBUG level in order to
set BASH debugging onto a function called `configureDebugLevel`. This
allows us to change the check without going thru all the scripts that
calls the check.