Add initial acceptance test suite
What is it?
This is a set of tests aimed to validate a deployed RabbitMQ release.

Why do we need it ?
We need it so that we can quickly validate a release prior to cutting it
when we introduce changes like adding a new erlang version or a new
rabbitmq version, just to a name a few.