Create dedicated diagnostics log and report files
Before this change, every time we deployed the release
and diagnostics were run, the output was appended to the
post-start.stdout.log and post-start.stderr.log files.

After this change, the diagnostics are captured on 2 seperate files
named as : diagnostics-<stamp>.log and diagnostics-<stamp>.report
The former captures stdout and stderr while executing the diagnostics
and the latter captures the diagnostics report with the list of tests
that passed and failed. Every deployment will produce this pair files.