RabbitMQ server job takes data snapshots before it stops
Diagnostics job consists of a list of test cases. Some of these test
cases compares live data with data that existed in RabbitMQ the last
time it was running. Example of this data are users, vhosts, etc.
The drain script of the rabbitmq-server job will take data snapshots
before it stops rabbitmq server.

The rabbitmq-server job also exposes a new property called `snapshots_directory`
which has the location of the snapshot file. The diagnostics job will
consume this property so that it can use this files.