post-start now waits for rabbitmq server to start
post-start used to run some checks which required rabbitmq server to be
available. However, Bosh was starting post-start before rabbitmq server
was ready. Rabbitmq server is started by the start script however the
post-start script was invoked even before the start script finished.
This was due to the fact that rabbitmq-server script (found in the
rabbitmq server package) created, very early, a pid file which made
monit/bosh to believe that it was ready when it was not.

_wait_for_rabbitmq-server waits until rabbitmq server process is fully
operational rather than waiting for its pid to be available.