Moved _take_snapshots to diagnostics package
Before this change, the script _take_snapshots was part of the
rabbitmq-server job. However, this script declared a variable that was
also referenced by the scripts in the diagnostic package. This variable
was $DIAGNOSTICS_SNAPSHOT and it had the location of the snapshots file.
This snapshot file contains a snapshot of rabbitmq server data such as
vhosts, users, etc.

After this change, the rabbitmq-server job depends on the diagnostics
package where we centralize all the scripts that relates to diagnostics
which are of two types: one that takes snapshots (_take_snapshots) and
the actual test cases (*_test.bash).

[#161617974]