Make prometheus_rabbitmq_exporter plugin available in v3.8
In order to make this work in 3.8, we had to change the list all
available plugins output. The script that we use to list plugins was
failing because the rabbitmq-plugins list output changed in 3.8. This
might not be intentional, but it's something that we had to address
before this would work. cc @hairyhum