Wait for Rabbit node to boot before running post-start
We need RabbitMQ to finish booting before we run pre-start commands. We
want to give it as much time as it needs, we simply don't know how long
it will take to recover state from disk and start all vhosts. It can be
seconds or it can be minutes, especially when there is a lot of data to
read from disk.