Make PerfTest run on deploy, as a regular job
Switched to the native binary, even though @acogoluegnes is recommending
against it: still experimental & performance not on par with JRE. At
least rabbitmq/rabbitmq-perf-test#208 is fixed in 2.9.0-rc.2. The goal
is to add back JRE support and allow this to run in both modes, but for
the purpose of this iteration, it is good enough.

PerfTest now automatically registers with Prometheus:
https://github.com/rabbitmq/rabbitmq-credentials/commit/ad4839b6640866d0b54ee4789d98c5883c6e7a48

This is the first job to make use of bpm:
https://bosh.io/docs/bpm/transitioning
It was far easier than expected, it made a lot of scripts redundant in a
nice way, the final job is really compact and simple.

Package tests are a first too. They helped catch a few unexpected
failures - worth the effort.

[#168539767]