Add support for setting which nodes PerfTest will connect to
We want to be able to force PerfTest to connect to a specific node so
that we get the best data locality when benchmarking. While this is not
realistic for real-world scenarios, it allows us to see what impact
spreading publishers & consumers across all nodes has on the overall
throughput & latency.