Remove rabbitmq-support-tools
This is a snapshot of the repo, so not the most up-to-date version, and
since we started moving a lot of this in rabbitmq-* commands, they are
less relevant than they used to. Ideally, we will have moved all
functionality from rabbitmq-support-tools into RabbitMQ itself, so we
can deprecate this for good. Small steps.