Add back the JVM-based PerfTest version, make it the default
We found the JVM-based PerfTest to be more performant, and it exposes
more metrics about the JVM, which are handy to have when trying to
answer questions like "Why does PerfTest appear blocked when RabbitMQ is
fine?"

It is still possible to use the GraalVM one, see the spec property
description.

There is a new `rabbitmq-perf-test` function available on shell login
(e.g. `sudo -i`). It wraps the PerfTest invocation configured in the
rabbitmq-perf-test job, defaults to:

  /var/vcap/jobs/rabbitmq-perf-test/packages/rabbitmq-perf-test-2.9/rabbitmq-perf-test-2.9.1/bin/runjava com.rabbitmq.perf.PerfTest

Part of this, updated PerfTest to latest stable, v2.9.1

cc @Vanlightly