Simplify Erlang version selection in add_erlang make target
We are only really interested in OTP 20+ patch releases. When we need to
ship RCs or first versions, add the OTHER option.

This is the behaviour that we want 99% of the time. Most often, we are
bumping patches, and it's difficult to see what we are going from-to
with the previous wall of version choices. Also, --version-sort FTW!