Make RabbitTestTool work on perftest instance
Add all topologies from
https://github.com/Vanlightly/RabbitTestTool/tree/master/deploy/topologies

Add rabbit-test-tool function wrapper that runs the point_to_point
topology. We don't want to invest more time into this as the Python
runner will be the one responsible for invoking RabbitTestTool. To test,
simply run `rabbit-test-tool` as root.

While iterating on this, we decided to use a glob to match all blobs
that should be part of the rabbitmq-perf-test-2.9 package. While
RabbitTestTool is not really PerfTest, because it's a fork, we want to
keep them together since we intend to merge them one day.

Signed-off-by: Jack Vanlightly <jvanlightly@pivotal.io>

[finishes #169981315]